+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ### Added
 - `[ir/def_tree.rs]`に`VarMetaData::is_mut`を追加
+- `[parse/mut_var.rs]`を追加
+- コンパイルエラーを追加
+- 不変変数と可変変数を追加
 
 ### Changed
 - `[ir/builder.rs]`のIncludeの処理を`[ir/builder/preproc.rs]`に移動